speachToText/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
openai-whisper>=1.0.0
requests>=2.26.0
ffmpeg-python>=0.2.0